Add recursive maze generator
The maze generation algo works flawlessly, however, I've had to include
some hacky code (i.e. export variable with walls instead of using
returned grid from function and useEffect to call the function). The main
reason is that the recursive function triggers the component to rerender
before i animate the walls - so this provides a separation of the generation
and output of the maze using useEffect hook and walls var respectively.

Additionally, I have tweaked init grid/start/target so the default recursive
maze renders with the perimeter filled in.